[baekjoon] "Silver" 문제들 채점 통과

+++ baekjoon/create-new.sh
@@ -1,5 +1,11 @@
 #! /bin/bash
 
-RUN="cp -r $1 ./$2 && cd ./$2"
-echo $RUN
-eval $RUN
+if [ -e "./$2" ]
+then
+    echo "$2 already exists."
+    cd "./$2";
+else
+    RUN="cp -r $1 ./$2 && cd ./$2"
+    echo $RUN
+    eval $RUN
+fi
